Update patch set 1
Patch Set 1: Code-Review+2

Looks OK.  I am pushing in order to get the basic file structures in place for the use case, but would ask that we review the file naming in the context of the SFC in it's entirety.  

This should be discussed on the SFC weekly call.   We should patch those file name changes (if necessary) once we have clarity on the potential need.

Patch-set: 1
Label: Code-Review=+2