Update patch set 1
Patch Set 1: Code-Review+2

As we are handing over due to Brady taking a break I will push this in. Howevere:
We need to get the MAC fix in from Yangtools or find a workaround!
Optimizations will be needed for node switch based ruleselection there is no major fault currently but we are filing rule tables un-neccesarrily with the current implementation.

Actions on the OF team to correct during next week.

Patch-set: 1
Label: Code-Review=+2