Update patch set 1
Patch Set 1: Verified+1 Code-Review+2

We will take care of Yang aspects discussed in the review when we merge the changes into the master (as they are still being discussed). This +2 for checking the changes into the "sfcofbasic" branch so that we can proceed with the next step of development.

Patch-set: 1
Reviewer: Gerrit User 2532 <2532@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=+1
Label: Code-Review=+2
Label: SUBM=+1