Update patch set 1
Patch Set 1:

Given the SF MAC & VLAN we can determine the egress interface of the the SFF can be externally discovered. The one that comes closest is the Host Tracker project. However, it assumes L2 learning to be enabled on the SFF.

We are proposing and approach where the NB can optionally provide the egress interface. This can be used for SFFs where such discoveries are not supported. We are looking at switches that are non OVS and purely OF based. Going through ODL is the only way to configure them. Here integrating such discoveries can get complex.

Again, it does not preclude discovery, as the interface entry in optional in Yang.

Patch-set: 1