Revert "added lisp updater for service functions and service function forwarders with lisp address"
This reverts commit 2fd4810db9dc9a22e472a40da37bf0cc91cee760.

Revert "added lisp updater for service functions and service function forwarders with lisp address"

This reverts commit 2fd4810db9dc9a22e472a40da37bf0cc91cee760.

Gal, David,
I am reverting this patch.

As noted in the comments to patch:
https://git.opendaylight.org/gerrit/#/c/10669/
This patch breaks the existing sfc-provider.

You were specifically asked not to do so in
the review for 10699 to which you provided no response.

The existing agreed architecture is extremely flexible and
allows for multiple independent providers, as you can see
in the openflow provider here:
	https://git.opendaylight.org/gerrit/#/c/10631/
	and here the provider for the ovsdb folks here:
	https://git.opendaylight.org/gerrit/gitweb?p=ovsdb.git;a=tree;f=ovs-sfc;h=8fa462c53b4c92b50a292ec3bc2a5b3e0ba22282;hb=HEAD
An additional provider from you guys that does something LISP
specific is very welcome, but breaking existing code the day
before code freeze is not.

Let me know if you need a hand getting your provider in and
a karaf feature for it.

Change-Id: Ie84ed0d4540af9e12ac20d55291871a133b7a793
Signed-off-by: Ed Warnicke <eaw@cisco.com>