Sfc-provider classifier functionality
+ classifier yang model change
+ classifier entry listener
+ basic API for ACL

How it works:
Whenever Classifier is attached to SFP it creates mapping between SFP <-> ACL.
Then, Classifier entry listenter calls method from ACL API which will update ACL
entries to point to the SFP.

If it comes to change of ACL in Classifier, old ACL entries are unset,
thus are not pointing to the SFP anymore. The new ACL is updated
appropriately.

Change-Id: Iba179fedee05756dbd3101242a481866ddfcba24
Signed-off-by: Andrej Kincel <akincel@cisco.com>