Update patch set 7
Patch Set 7:

What are the next steps here?

This patch is causing the merge job to fail with Sonar unable to process the random characters created by RandomStringUtils. I found the documentation at [1] but it does not mention if it's returning valid UTF-8 strings or not. But I suspect it isn't and that's likely why Sonar's not able to process the string data. The root pom.xml file for this project specifies that sourceEncoding should be UTF-8 so if RandomStringUtils is not returning UTF-8 compatible string data then the test would need to be modified to get string data from a UTF-8 compatible function.

[1] http://commons.apache.org/proper/commons-lang/apidocs/org/apache/commons/lang3/RandomStringUtils.html

Patch-set: 7
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0