Update patch set 2
Patch Set 2:

thanks Andrej.

I think instead of traversing all SCL and sending the ACL to all of them (even the ones that do not need it), we need to attach the listener only to the classifier tree. 

If a classifier is committed we get the ACL (we have its name) and send only to the SFF found in the classifier. 

I think we should also have a discussion on whether we should attach listener on top of tree or at the element level. The original listeners are all attached to the element level so there is never a  need to traverse anything.

Patch-set: 2
Reviewer: Gerrit User 248 <248@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0