Update patch set 2
Patch Set 2:

Yes, you are right. It's better to listen on SCL changes. The problem is, that after ACL name is set in SCL, it's not guaranteed that ACL actions will be set to appropriate RSP (based on RSP from SCL) before sending the ACL. 

In simplier form: by changing SCL, you are triggering changes in ACL. You have to wait for the changes to apply and only after that you can send the ACL to SFFs from SCL. That's why my first approach was to listen on ACL changes.

Patch-set: 2
Reviewer: Gerrit User 2138 <2138@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0