Update patch set 2
Patch Set 2:

I do not think we should do a more convoluted code in order to prevent configuration screw ups by the user. The user needs to do the minimum.

Also, looking at the code, it seems we are ending the ACLs to all SFF, irrespective if their classifier access list name matches the name on the onDataChanged() callback.

Patch-set: 2