Update patch set 10
Patch Set 10:

Where are you reading that classifier is just informational?

The official Maven documentation states that it is for distinguishing multiple artifacts that were produced from the same pom.xml file. Specific quote here:

---- SNIPPET ----
classifier:
The classifier allows to distinguish artifacts that were built from the same POM but differ in their content. It is some optional and arbitrary string that - if present - is appended to the artifact name just after the version number.
As a motivation for this element, consider for example a project that offers an artifact targeting JRE 1.5 but at the same time also an artifact that still supports JRE 1.4. The first artifact could be equipped with the classifier jdk15 and the second one with jdk14 such that clients can choose which one to use.

Another common use case for classifiers is the need to attach secondary artifacts to the project's main artifact. If you browse the Maven central repository, you will notice that the classifiers sources and javadoc are used to deploy the project source code and API docs along with the packaged class files.
---- SNIPPET ----

You can see the reference on this page:

http://maven.apache.org/pom.html#POM_Relationships

Patch-set: 10