Update patch set 10
Patch Set 10:

> I noticed that southbound-impl has a file called maven-metadata-local.xml
 > which seems to classify the artifact as config.

This is used by Maven Artifact Resolver to determine if it should check with an upstream server again for updates to files and redownload if necessary. It's not data used by a build to determine if a file is a dependency or not so Maven Compiler / Karaf does not use this file to make any determination what dependencies a bundle has.

Patch-set: 10