Rewrite the service funtions scheduling API
This patch rewrites the service funtion scheduling API.
The origin version of API is defined in:
https://git.opendaylight.org/gerrit/#/c/14118/
It defines an API for scheduling a single service function.
Some algorithms may schdule service functions for the whole
chain. New API schedules SFs for the whole chain.

Change log:
Solve the conflicts with commit:
https://git.opendaylight.org/gerrit/#/c/16016/1
SFF locator name is filled in corresponding RSP leaf

Signed-off-by: Johnson Li <johnson.li@intel.com>