Changing how RenderedServicePathHop List is created
Im working on populating the RSP Hop ServiceFunctionForwarderLocator,
and the recently added funcionality made that difficult.

Additionally, it doesnt make sense for the RSP Hop list
to be created in the Service Function Scheduling code.
Previously, if the SF scheduler didnt find the SFs, then
the RSP Hop list was created in SfcProviderRenderedPathAPI,
meaning that the RSP Hop list creation code was duplicated
in the SF schedulers and in SfcProviderRenderedPathAPI.
Now, the RSP Hop list creation is only in
SfcProviderRenderedPathAPI, thus increasing the SF
scheduler classes cohesion.

A separate patch is coming soon to populate the RSP Hop ServiceFunctionForwarderLocator

Change-Id: I17905ca4b652528876e92341ace8e05fb3601e24
Signed-off-by: Brady Johnson <brady.allen.johnson@ericsson.com>