Update patch set 1
Patch Set 1: Code-Review-1

Need unit tests, javadocs, etc, etc. 

You created a SFCProviderGrapth but there is already one in sfcofl2 package. It seems to me we shoudl consolidate and move this code to a common place outside provider package.

the changes to Yang were not clear to me. The description needs to be improved considerably.

Patch-set: 1
Label: Code-Review=-1