Update patch set 1
Patch Set 1:

> Very interesting. How can this be actually used? 

With OVS : I was able to build and use OVS as a SC and a SFF with Pritesh's NSH patch.(BTW, I already created a pull request to add the compatibility with kernel 3.16) 
I use scapy to emulate a service function : it decreases the service index and reemits the packet.

 > Can you maybe
 > include some examples in the file so others can also test it?

Of course: I'll add examples (and my emulated SF script with scapy ) in the file as soon as possible.

Patch-set: 1
Reviewer: Gerrit User 3640 <3640@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0