Update patch set 1
Patch Set 1:

In retrospect I wish we had more discussion about this. The idea is good but I think we need a different approach. What is video? Is video in qosmos the same as Video in Ericsson or Cisco? How about P2P? Interoperability would be impossible.

I suggest something like this:

https://tools.ietf.org/html/rfc6759

Patch-set: 1