Update patch set 1
Patch Set 1:

> In retrospect I wish we had more discussion about this. The idea is
 > good but I think we need a different approach. What is video? Is
 > video in qosmos the same as Video in Ericsson or Cisco? How about
 > P2P? Interoperability would be impossible.
 > 
 > I suggest something like this:
 > 
 > https://tools.ietf.org/html/rfc6759

It's exactly the idea.
I have another patch in the pipe which contains another yang model augmentation: it contains 1 dictionary for each uint32 to human readable values.
Thanks to this dict, we will be able to load and update the IPFIX dictionary, for each L7 matcher, so we don't have to focus on the semantic but only on the function.

What about a call to discuss about this ? I'll send you an email.

Patch-set: 1
Reviewer: Gerrit User 3640 <3640@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0