Update patch set 1
Patch Set 1:

> (1 comment)

Reinaldo, thanks for the comment.
Yet, I have 2 questions :
- by defining the type 'enum class-id-name', should'nt we delete the container class-id-dictionary and enforce the type of application-id/class-id to this enum ?
But doing this way, we can not add a new class-id at runtime thru JSON API, and from what I understand, values from 21 to 255 are available.

If we keep both, I'm afraid to disturb the final user (should I use the enum or values present in the dictionary ?)

I suggest to keep the dictionary as is but to provide default values thru bootstrap json configuration.

- you suggest to modify the type selector-id to application-id.
Shouldn't we keep a type named 'selector-id' and inherit from it to create an other type named 'application-id' ?
I understand within IPFIX RFC, the name "selector ID" is prefered to have more genericity. 
In our case, if we use the class-id 'USER-DEFINED', the name 'application-id' is not relevant.

A use case would be the subscriber-class : according to this value (provided by the operator, not the classifier) we may prioritize the flow.

Patch-set: 1
Reviewer: Gerrit User 3640 <3640@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0