Update patch set 1
Patch Set 1:

> >
 > > Reinaldo, thanks for the comment.
 > > Yet, I have 2 questions :
 > > - by defining the type 'enum class-id-name', should'nt we delete
 > > the container class-id-dictionary and enforce the type of
 > > application-id/class-id to this enum ?
 > > But doing this way, we can not add a new class-id at runtime thru
 > > JSON API, and from what I understand, values from 21 to 255 are
 > > available.
 > 
 > I do not think there is a use-case for adding a new one at runtime
 > since you can use "User-defined" for anything
 > 

If I understand, the dictionary will be populated with only values 
from the enum. What's the point of the dictionary in this case ?
We should directly refer in the application-id to the value in the
 enum and delete the dict.

But that's not what I would like :
I agree the class-id-dictionary will not be modified often.
Yet, to give the user the possibility to add it's own class-id 
(from 21 to 255) is a plus :
of course we can use the "User-Defined" class, but if the user 
already has multiples criteria categories, it will force the 
user to split the selector-id field and to manage the bit field 
manually thru a lookup table or a global configuration file.


Another usecase would be to chain 2 classifiers in the same system:
 1 open source and 1 proprietary, we would need 2 differents class-id.
In this case, it will be very hard to map everything within the same class-id.


From my point of view, as we can offer this possibility without 
transgressing the RFC, we should.
From the RFC point of view, if an existing implementation (all 
reserved class-id) wishes to use their existing class-id, they 
must rebuild ODL to add their own value in the yang model.

From the UI, we will be able to display and filter the different
class-ids, while if we just use the USER-Defined category, we will
have to refer to some lookup table: Not easy to use IMHO.

 > >
 > > If we keep both, I'm afraid to disturb the final user (should I
 > use
 > > the enum or values present in the dictionary ?)
 > 
 > 
 > The dictionary refers to the enum. I fixed the values in the enum
 > if that was the issue.
 > 

Nope, I modified the values according to the RFC when I built the model :-)

 > 
 > >
 > > I suggest to keep the dictionary as is but to provide default
 > > values thru bootstrap json configuration.
 > >
 > > - you suggest to modify the type selector-id to application-id.
 > > Shouldn't we keep a type named 'selector-id' and inherit from it
 > to
 > > create an other type named 'application-id' ?
 > 
 > 
 > your container  was class-id + selector-id, that is defined as
 > application-id, not selector-id.

Ok, I missunderstood the RFC, and made a shortcut.

 > 
 > If you want to create yet another container called selector-id that
 > is fine.
 > 
 > 
 > > I understand within IPFIX RFC, the name "selector ID" is prefered
 > > to have more genericity.
 > 
 > 
 > selector id is not the same as application id. This is defined in
 > section 4.
 > 
 > application-id = class-id + selector+id
 > 
 > selector-id = [PEN] + unit32
 > 
 > 
 > 
 > > In our case, if we use the lass-id 'USER-DEFINED', the name
 > > 'application-id' is not relevant.
 > 
 > I believe it does.
 > 
 > >
 > > A use case would be the subscriber-class : according to this
 > value
 > > (provided by the operator, not the classifier) we may prioritize
 > > the flow.
 > 
 > 
 > I start to see your use-case. But at some point you still need to
 > define an application-id that maps to all network flows to/from
 > this subscriber otherwise the classifier.
 > 

Yes, that's why I supposed it would be great to be able to create
a class-id, and then specific application id based on this class-id, while not messing with other categories.

Patch-set: 1