Create patch set 2
Patch-set: 2
Subject: 1.Description This project is branch of service function chain project belonging to OpenDaylight (https://wiki.opendaylight.org/view/Service_Function_Chaining:Main).  We add a self-defined greeady algorithm called Two-Level-Optimization in master code of service function chain.  The algorithm implements optimized service chain routing, with the goal of maximum throughput.  Every time when a customer arrives, the step of resource-check is wakened to judge whether there are enough resources to serve the customer.  If the resource is insufficient, a new virtual machine is created and powered on.  In daisy-chain network connection mode, the issue of path retracing might happen, which leads to the waste of link bandwidth.  Two-Level-Optimization algorithm solves the problem by introducing a mechanism to prevent path retracing.  Besides, every time when choosing a service chain path, the algorithm chooses the one with the ability to serve the most customers in order to evenly utilize the resource of virtual machines.
Commit: dbbeda3e292ea3b0ca12f47f26fb8f4ed10dd912