Changed typing of SFC-OVS Options.
The ovs-options were strictly typed. Unfortunately,
the options in OVS are key/value pairs of strings,
and strict typing prevents the use of values like
"flow" for remote_ip and other options. This patch
converts these types to strings so they can support
"flow" values and can be passed directly to the
OVSDB Southbound.

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>