Separating config files of features
Features which are not included in odl-sfc-provider feature
(odl-sfc-test-consumer, odl-sfc-netconf and odl-sfc-sb-rest) should not
be listed as required capabilities in 80-sfc.xml. This is causing config
subsystem exception, if only odl-sfc-provider feature is installed and
the above listed features are not.

The behaviour should be:
1. odl-sfc-provider can be installed without any other features except
odl-sfc-model
2. a SFC renderer (sfc-ovs, sfcofl2, sfc-netconf) is dependent on
odl-sfc-provider, thus installation of renderer feature will triger
odl-sfc-provider installation too.

+ separating sfc-sb-rest config file
+ separating sfc-netconf config file
+ separating sfc-test-consumer config file
+ removing sfc-test-consumer from boot features

Change-Id: I6cd942dfb43abfa5728475524c3c5be869bdc2f1
Signed-off-by: Andrej Kincel <akincel@cisco.com>