Update patch set 3
Patch Set 3:

I guess in my mind it's a question of what we think the architecture should be -- do we want the renderers to be separate/modular from the provider? That's probably a bigger question than this gerrit, but since this moves further in that direction, it seemed like a good place to start.

Patch-set: 3