Update patch set 1
Patch Set 1:

Yi... I am trying to trim down sfc-provider to be leaner and sfc overall to be more modular.

l2switch-hosttracker is kind of heavy to be something that *has* to be included in the core of SFC whether or not we even have OF switches in the mix.

I absolutely have *no* issue with being able to do something like this, but am trying to trim down the *core* SFC dependencies and make things modular enough so it can be as light as needed for whatever the particular use case is.

Patch-set: 1