Update patch set 2
Patch Set 2: Code-Review+1

Reinaldo, we are using sfclisp, and plan to extend it in Be. Great point regarding test coverage, we'll work on this. Regarding the breakage, it has happened only once before, and the fix is submitted minutes after the breaking patch. Moving forward, to further minimize disruption in such cases, we will make sure that the fix is pushed to sfc in advance, and we'll notify the sfc-dev mailer ahead of time. I hope this addresses your concerns. Thanks.

Patch-set: 2
Reviewer: Gerrit User 654 <654@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1