Update patch set 9
Patch Set 9:

One question. The way the SFC-UI would pick up new Yang changes automatically and reflect on the UI was through the Yin files. For example, if I changed a Yang leaf name anf generated a new Yin file, the UI would automatically change. This was very handy.

But you removed the directory, so would we proceed to do that now?

Patch-set: 9