Update patch set 4
Patch Set 4: Code-Review-1

I'd prefer we didn't merge this until we see the implementation impacts on SFCOFL2. I don't think we should add functionality that isn't consumed anywhere. Until someone makes an implementation change that supports the use-case, or has immediate plans to implement it. Can we keep this unmerged and when someone decides to implement, make it a dependent change?

Patch-set: 4