Update patch set 5
Patch Set 5:

These are my previous comments. They are not related to current patch, adding here for continuity so we do not need to keep discussing the same things over and over

"If you use references such as service-function-forwarder-ref instead of absolute paths then SFF model can be changed (or even reved) without having to remember and fix all models that have the path hardcoded.

Also, importing SF into SFF will create circular dependency in the future if you ever try to add a reference to SF in SFF. The way to break that is create a separate module which is imported by both SF and SFF. We did such things for locator, types, etc.

Patch-set: 5