Update patch set 5
Patch Set 5: -Code-Review

Reinaldo, this patch does exactly that, uses leafrefs in the module itself with a path so that its not referenced explicitly in another module. Can you make a comment on the actual patch loc where this path is that shouldnt  be?

On your second point, i whole heartedly agree. This is the purpose of the mapping container in a separate module and my intention was to remove ALL traces of sf references in sff and all traces off sff in sf, when i did the impl as this is a MAJOR issue  and always has been. 

So im assuming you you will be ok with me going ahead and removing those redundant cross references to sf in sff and vice versa now and could you point explicitly in the patch where my leafref is wrong and i can get on with the fix.

Patch-set: 5