Update patch set 9
Patch Set 9:

Because not every form in SFC UI is Yang driven, you still need to contact us to check if it needs to be changed.
We will connect SFC UI to our version of YangUtils (because in the SFC is still the fork). After that change it will read Yin from controller not from the file.

Patch-set: 9