Update patch set 19
Patch Set 19:

(1 comment)

The previous patch used path reference for certain SF and SFF names. This patch uses a typedef that is in fact a string. This means the enforcement from Yang is gone. Will the path reference be brought back in a subsequent patch?

As far as all the other changes they are fallout product from the introduction of the common model. Since new model changes SF and SFF and those are are used everywhere, many changes are needed.

Patch-set: 19