Update patch set 19
Patch Set 19:

Thanks Reinaldo, I should have mentioned in the commit message or an email but instead I mentioned on the SFC call. I had originally intended to do both typedefs and leafrefs. The issue was as I started to apply them consistently I ended up getting cyclical dependencies. Partially due to file structure, partially due to the model. The follow on from this patch would be a revisit of aspects of the model, which may change where we insert referential integrity. So I thought it best to address that as a requirement in any further model changes. That being said, if there's areas that absolutely still need it in the interim, I'm happy to work those in. Would it be acceptable that in the interim they are inline? Ie no leafref typedefs and we move to a consistent pattern once we get a lot of the model worked out?

Re: IETF specs. I can see how we need to be compliant if we are going to publish these as actual specs. Sfc-common was intended to be just a convenience library, never published as a spec. Although I can see if we go down that road, it may force each yangfile to be independent. For style purposes if you'd prefer the same format, I can reformat and resubmit. Let me know, happy to make changes, but prefer to address both leafrefs and formatting in one patchset.

Patch-set: 19