Update patch set 1
Patch Set 1:

For the first time and after the Yang model change. We will connect SFC UI to the DLUX version of Yangutils and it will be not necessary then.

This was in the README in the sfc-ui:

After changing YANG models it is required to execute (requires pyang-1.5 to be installed and placed on $PATH):
sh
$ ./convert_yang_to_xml.sh

This will generate XMLs from YANG model needed for dynamic UI generation. In future, this would be not more necessary
because XMLs will be also available from ODL controller via RESTconf (URL: restconf/modules/)

Then build project:
$ mvn clean install

Patch-set: 1
Reviewer: Gerrit User 3799 <3799@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0