Update patch set 5
Patch Set 5:

thanks for the comments. It was not the intention to be merged, just to foster discussion. I should have out [draft] in front of it.  This is Keith's idea and he will work on it.

Patch-set: 5
Reviewer: Gerrit User 248 <248@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0