Update patch set 4
Patch Set 4:

Here is the net problem I am working towards this morning on this patch. Just so we clear, I'm not looking for help just adding back dependencies to pass Jenkins, quite the opposite. Here's the issue so its all documented in the patch itself. The original PS was the MVP for POMs for the running code. ie not UT. When you add back in UT, not only are extra dependencies needed but also some of those INDIVIDUAL Unit Tests (not classes, actual tests) take MINUTES to run locally and in Jenkins. So rather than chucking dependencies back in, which I started to do to see if I could find out using Jenkins instead of my local machine what it takes to run with UT, I'm going to focus on the intent of this patch, removing as many external dependencies as possible (as simple as possible but no simpler). This will make things much cleaner, and less prone to all the external breakages SFC seems more prone to than any other project. So sure, this may look like simply going through and deleting stuff from POMs that were copypasta'd, which is normal, but what I am trying to do with this is dramatically increase stability by removing unnecessary dependencies for the RUNTIME code, and centralise versions and params for consistency etc.

So small patches == good, I get it... but also note that if I do this as two separate patches the full intent will be lost. Making the whole package a lot simpler.

-1'ing patch so that folks are aware I'm working on it, and not to bother reviewing until I remove the -1.

Patch-set: 4