Update patch set 5
Patch Set 5:

When initialization flows are removed in the current implementation, I should also mark them as not-initialized. Besides initialization of SFFs is called every time RSP is created (not only after first RSP) so a bug should not be nested by this commit. I guess that it is not an extraordinary use-case to clear paths with SFFs and create new configuration without restarting ODL. Now it's not possible to do this and I have to know implementation details if I want to work around the problem.

Patch-set: 5
Reviewer: Gerrit User 3547 <3547@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0