Update patch set 5
Patch Set 5:

To be more specific about the problem: Init flows are removed from SFFs when the last RSP is removed. After init flows are removed from a SFF the internal state for the SFF still holds "is initialized" value. The fix is to change the state of the SFF to "not initialized" so it's possible to create init flows when first RSP using SFF is created. This bug fix actually solves the bug in the current implementation.

Patch-set: 5