Update patch set 14
Patch Set 14:

There is a problem with this patch regarding using sfc-common. In order to truly reap the benefits of integrity the names should be like

"name": "sfc-common:SFC1"

But the UI, Python regression tests and the datastore is not stripping the "sfc-common" prefix. So, we need to use sfc-common through RESTconf but the UI needs to strip it for the user. At the same time the UI needs to add sfc-common under the covers when a user configures new function, forwarder, chain, etc. 

I'm fixing the Python regression (sfc-test) tests but much more needs to be done. We also need somebody to fix the UI otherwise there will confusing and certain things will not work anymore.

For example the UI RSP RPC is no adding the sfc-common automatically so it is not working anymore. You get a ""Service Function Path does not exist" because now the path name needs to have "sfc-common" prefixed.

Patch-set: 14