Update patch set 1
Patch Set 1: Code-Review-2

I do not follow some of the rationale, although I understand the oper/config patterns, which btw is still undergoing quite a lot of discussions on the IETF : https://tools.ietf.org/html/draft-openconfig-netmod-opstate-01

If something exists only in ODL and you want to make it oper then insert a "config false" and you are done. Some of these containers purpose is to track who is using what and do not reflect operational state found in a device. In this case the pattern does not apply.

OVSDB uses the oper/config thing because it talks to devices directly, we do not, we talk to OVSDB (or Netconf). So we have second-hand view of things. 

I think deciding how this pattern applies to SFC needs a broader discussion instead of blindly following it.

Patch-set: 1
Reviewer: Gerrit User 248 <248@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2