Update patch set 2
Patch Set 2:

Yes, I disagree (for now) with the oper approach. We should evaluate all proposals from IETF and figure out which one is going to be adopted. All of them try to solve the problem of intent vs. actually configured besides other things. I contacted some folks and am waiting for feedback.

It is easier to make a case in things like netconf but not ovsdb. OVSDB has its own datastore, models and so forth. What SFC gets is OVSDB's view of the device, not what is actually configured in the device. So, it is a oper "twice removed"

You can take a look at the drafts and see the competing proposals to solve these things. 

Finally you also deprecated the RPC to delete RSPs. I'm not sure how user will be able to delete RSPs.

Patch-set: 2