Update patch set 1
Patch Set 1:

> SF acts a role as a router does, so it is inappropriate to swap IP
 > addresses, I think you can add an option to switch on/off this if
 > it is really required by your user scenario.
 > 
 > In most cases (especially for firewall), you can't change IP header
 > except TTL.

We should sort this out.  This code is emulating the SF behavior, and I think that this is how an SF would really work.  Consider two cases:

Case 1:  this code uses a raw socket, but a real SF that works with VXLAN-gpe would likely open a UDP socket for receiving and sending.  With a typical UDP socket, the IP SA and DA would end up getting swapped on rx and tx.

Case 2:  The SF is not directly attached to the SFF.  In this case, NSH/VXLAN-GPE would need to be routed to the SF and then back to the SFF.  On the SFF->SF path, the DA=SF, and the SA=SFF.  On the SF->SFF path, the DA=SFF and the SA=SF.

Note: we're not changing the IP header of the inner packet -- just the tunnel header to send the packet back to the SFF.

So, I'm happy to add an option, but I think this should be the default behavior.

Patch-set: 1