Update patch set 1
Patch Set 1:

Hi Yi,
The modification of the tunnel dst/src IP is correct.  This is not a router, because the tunnel destination is the VNF itself.  Since the packet is destined to him, he consumes the packet.

If the packet was destined to another IP but was routed to him, then he would responsible for only routing.

AFAIK in the ODL SFC setup, you give the destination tunnel location (DPL for SF) you are telling SFC that SF Is the destination IP for the tunnel, which means the SF should consume it.

Patch-set: 1
Reviewer: Gerrit User 5089 <5089@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0