Add NSH proxy into SFC Yang model.
Yang model in SFC by default assume all SFs are NSH-aware,
but it needs to be extended to support NSH-unaware SF which is not covered in current Yang model.
With the NSH-uaware SF, a NSH proxy needs to be added in the model that sits between SFF and NSH-unaware SFs.

Change-Id: Id4705e0e430ba2f7a50bc2b9fd29d3c2ddb24fa2
Signed-off-by: Hongjun Ni <hongjun.ni@intel.com>