Add NSH proxy into SFC Yang model
Yang model in SFC by default assume all SFs are NSH-aware,
but it needs to be extended to support NSH-unaware SF which is not covered in current Yang model.
With the NSH-uaware SF, a NSH proxy needs to be added in the model that sits between SFF and NSH-unaware SFs.

Change-Id: Ie8c1a3a3acdb16d23be0045b2dd970292438d55e
Signed-off-by: Hongjun Ni <hongjun.ni@intel.com>