Update patch set 3
Patch Set 3: Code-Review-2

No changes are allowed to the API now that we've passed the M4 milestone. If this change is a bug, then you'll need to create a Bugzilla and send an email to the sfc-dev list.

Patch-set: 3
Reviewer: Gerrit User 2039 <2039@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2