Update patch set 3
Patch Set 3:

> Not against the change per se but it seems like you could therefore
 > scrap a lot of sfc-parent/pom.xml too right?

Yes, all the external dependency stuff in sfc-parent can go (the only dependency that was left is slf4j-simple, and that can be removed too; I'll update this patch). The plugin stuff can probably be cleaned up as well but that requires a bit more care; I'm driving this set of changes from the odlparent side of things, I'm queuing up project reviews for later :-). The main aim of these patches is to get rid of properties defined in odlparent, because they're a pain to manage (and plain dependency management is designed for that).

 > What is the point of in-project dependency management if its all in
 > odlparent?

You still need in-project dependency management if you want to centralise your intra-ODL dependencies, as is done in sfc-parent; I don't think it would be good for ODL to start defining these in odlparent (in fact I'd much rather decouple projects). In-project dependency management is also useful for cases where you want a consistent dependency for your project that's not defined in odlparent, which IMHO is perfectly allowable. I don't want to get into a situation where projects aren't allowed to introduce a new external dependency if it hasn't been vetted by the odlparent Powers That Be... But I do still want to be able to pull dependencies up to odlparent when it makes sense to do so, and drive upgrades from there (which is good for you because it puts the onus on me to manage the fallout).

 > Is there any issue with at some point ODL parent
 > branching stable/foo while we are on master?

The odlparent you get is determined by the version you pull in (1.6.0-SNAPSHOT currently), so when odlparent master is branched to stable/be, you'll start pulling that in, until you update your master to pull in whatever the new version of odlparent master is. I don't think there should be any issue with that, either on your master branch or your stable/be branch; in any case, I try to take care not to break things in projects when I make changes in odlparent, and if you see any breakage I'll do whatever's necessary to fix it (either in your project or by reverting odlparent). There will be a more detailed versioning proposal soon that should put all these issues to rest (assuming the community agrees with it!).

Patch-set: 3
Label: Verified=0