Update patch set 6
Patch Set 6:

Apologies for confusion. I was actually still using GBP MASTER artifacts when I thought this was working. The culprit I believe to the failure is this:
 table=1, n_packets=0, n_bytes=0, priority=250,udp,tp_dst=6633 actions=goto_table:4

Tunnels are decapped so the tp_dst will be the transport DestPort of the inner packet, not the outer tunnel. 

You should use the OFPORT of the tunnel interface, but not "INPORT". This way you have more control over which particular VXLAN GPE tunnel you use, especially in the case where potentially we make use of Push/Pop with Kernel LWT for VXLAN GPE/NSH encap and have multiple tunnels.

This can be kept in a map potentially. See SwitchManager in GBP OfOverlay for some ideas on how this could be done... but note we use a portName heuristic, but since this is done explicitly in the SFF config via REST, you can use that ... and SFCOVS should be of some help here too.

Patch-set: 6