Update patch set 4
Patch Set 4: Code-Review-1

GBP SFC Demo works but 
curl -i -H "Content-Type: application/json" --data '{"sfc-of-renderer-config" : {"sfc-of-table-offset" : 20, "sfc-of-app-egress-table-offset" : 80}}' -X PUT --user admin:admin http://localhost:8181/restconf/config/sfc-of-renderer:sfc-of-renderer-config/

has no effect on existing flows. 

Using the above before configuring flows works but would prefer:
goto_table:80
be changed to resubmit:(,80)

This way later on we can have the option to add metadata and not assume a linear pipeline. 

There should also be a listener on this datastore that reacts to changes, otherwise you need to restart the controller to change it.

Patch-set: 4
Label: Code-Review=-1